less use of type "long" in pyx files (outside of p-adics)

+++ src/sage/libs/ntl/ntl_lzz_p.pyx
@@ -139,7 +139,7 @@ cdef class ntl_zz_p():
             self.c = <ntl_zz_pContext_class>modulus
         elif isinstance(modulus, Integer):
             self.c = <ntl_zz_pContext_class>ntl_zz_pContext(modulus)
-        elif isinstance(modulus, long):
+        elif isinstance(modulus, int):
             self.c = <ntl_zz_pContext_class>ntl_zz_pContext(modulus)
         else:
             try:
